+++ src/components/data-hooks.js
@@ -385,3 +385,60 @@ export function useRasterData(loaders, dataset, setItemIsReady, addUrl, isRequir
 
   return [raster, imageLayerLoaders, imageLayerMeta];
 }
+
+/**
+ * Get data from a genomic-profiles data type loader,
+ * updating "ready" and URL state appropriately.
+ * Throw warnings if the data is marked as required.
+ * Subscribe to loader updates.
+ * @param {object} loaders The object mapping
+ * datasets and data types to loader instances.
+ * @param {string} dataset The key for a dataset,
+ * used to identify which loader to use.
+ * @param {function} setItemIsReady A function to call
+ * when done loading.
+ * @param {function} addUrl A function to call to update
+ * the URL list.
+ * @param {boolean} isRequired Should a warning be thrown if
+ * loading is unsuccessful?
+ * @param {(function|null)} onLoad An extra function to execute upon load of the data.
+ * @returns {array} [neighborhoods] where
+ * neighborhoods is an object.
+ */
+export function useGenomicProfilesData(
+  loaders, dataset, setItemIsReady, addUrl, isRequired, onLoad = null,
+) {
+  const [genomicProfilesAttrs, setGenomicProfilesAttrs] = useState();
+
+  const setWarning = useSetWarning();
+
+  useEffect(() => {
+    if (!loaders[dataset]) {
+      return;
+    }
+
+    if (loaders[dataset].loaders['genomic-profiles']) {
+      loaders[dataset].loaders['genomic-profiles'].load().catch(e => warn(e, setWarning))
+        .then((payload) => {
+          if (!payload) return;
+          const { data, url } = payload;
+          setGenomicProfilesAttrs(data);
+          addUrl(url);
+          if (onLoad) {
+            onLoad();
+          }
+          setItemIsReady('genomic-profiles');
+        });
+    } else {
+      setGenomicProfilesAttrs(null);
+      if (isRequired) {
+        warn(new LoaderNotFoundError(dataset, 'genomic-profiles', null, null), setWarning);
+      } else {
+        setItemIsReady('genomic-profiles');
+      }
+    }
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [loaders, dataset]);
+
+  return [genomicProfilesAttrs];
+}